Rename package p to util
Work with `p` as a package name was not intuitive. Better use util for now.
